requirements.txt
jsonpath==0.82
loguru==0.7.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
